+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow type-checking specs to be created by passing a type directly to `s` (#12)
 - Add email address string format (#6)
 - Add URL string format factory `s.url` (#16)
+- Add Python `dateutil` support for parsing dates (#5)
 
 ### Fixed
 - Guard against `inspect.signature` failures with Python builtins (#18)